+++ packages/headless/src/controllers/product-listing/facet/headless-product-listing-facet-manager.ts
@@ -2,6 +2,7 @@ import {facetOptions, productListing} from '../../../app/reducers';
 import {loadReducerError} from '../../../utils/errors';
 import {sortFacets} from '../../../utils/facet-utils';
 import {buildController} from '../../controller/headless-controller';
+import {CoreFacetManager} from '../../core/facet-manager/headless-core-facet-manager';
 import {
   FacetManager,
   FacetManagerPayload,
@@ -10,7 +11,12 @@ import {
 import {ProductListingSection} from '../../../state/state-sections';
 import {ProductListingEngine} from '../../../app/product-listing-engine/product-listing-engine';
 
-export type {FacetManager, FacetManagerState, FacetManagerPayload};
+export type {
+  CoreFacetManager,
+  FacetManager,
+  FacetManagerState,
+  FacetManagerPayload,
+};
 
 /**
  * Creates a `FacetManager` instance for the product listing.