+++ packages/quantic/force-app/main/default/lwc/quanticInsightInterface/quanticInsightInterface.js
@@ -1,0 +1,53 @@
+import {LightningElement, api} from 'lwc';
+import {
+  getHeadlessBindings,
+  loadDependencies,
+  setEngineOptions,
+  HeadlessBundleNames,
+} from 'c/quanticHeadlessLoader';
+
+// @ts-ignore
+import getHeadlessConfiguration from '@salesforce/apex/InsightController.getHeadlessConfiguration';
+
+/** @typedef {import("coveo").InsightEngine} InsightEngine */
+/** @typedef {import("coveo").InsightEngineOptions} InsightEngineOptions */
+
+export default class QuanticInsightInterface extends LightningElement {
+  /**
+   * @api
+   * @type {string}
+   */
+  @api engineId;
+  /**
+   * @api
+   * @type {string}
+   */
+  @api insightId;
+
+  /** @type {InsightEngineOptions} */
+  engineOptions;
+
+  connectedCallback() {
+    loadDependencies(this, HeadlessBundleNames.insight).then(() => {
+      if (!getHeadlessBindings(this.engineId)?.engine) {
+        getHeadlessConfiguration().then((data) => {
+          if (data) {
+            this.engineOptions = {
+              configuration: {
+                ...JSON.parse(data),
+                insightId: this.insightId,
+              },
+            };
+            setEngineOptions(
+              this.engineOptions,
+              CoveoHeadlessInsight.buildInsightEngine,
+              this.engineId,
+              this,
+              CoveoHeadlessInsight
+            );
+          }
+        });
+      }
+    });
+  }
+}