+++ scripts/packages.js
@@ -1,0 +1,17 @@
+const packageDirsNpmTag = [
+  'atomic',
+  'auth',
+  'bueno',
+  'headless',
+  'atomic-react',
+  'atomic-angular/projects/atomic-angular',
+  'quantic'
+];
+
+const packageDirsSnyk = [
+  'headless',
+  'atomic',
+];
+
+exports.packageDirsNpmTag = packageDirsNpmTag;
+exports.packageDirsSnyk = packageDirsSnyk;